+++ src/main/java/com/shanebeestudios/skbee/elements/other/OtherElementRegistration.java
@@ -17,6 +17,7 @@ import com.shanebeestudios.skbee.elements.other.effects.EffBlockRandomlyTick;
 import com.shanebeestudios.skbee.elements.other.effects.EffBlockstateUpdate;
 import com.shanebeestudios.skbee.elements.other.effects.EffBreakBlocksWithEffects;
 import com.shanebeestudios.skbee.elements.other.effects.EffChunkRefresh;
+import com.shanebeestudios.skbee.elements.other.effects.EffChunkUnload;
 import com.shanebeestudios.skbee.elements.other.effects.EffDispatchCommand;
 import com.shanebeestudios.skbee.elements.other.effects.EffDropItem;
 import com.shanebeestudios.skbee.elements.other.effects.EffEntityBlockStorage;
@@ -26,7 +27,7 @@ import com.shanebeestudios.skbee.elements.other.effects.EffEquipmentChange;
 import com.shanebeestudios.skbee.elements.other.effects.EffFreezePlayerConnection;
 import com.shanebeestudios.skbee.elements.other.effects.EffGiveOrDrop;
 import com.shanebeestudios.skbee.elements.other.effects.EffHurtAnimation;
-import com.shanebeestudios.skbee.elements.other.effects.EffLoadChunk;
+import com.shanebeestudios.skbee.elements.other.effects.EffChunkLoad;
 import com.shanebeestudios.skbee.elements.other.effects.EffOpenContainerAnimation;
 import com.shanebeestudios.skbee.elements.other.effects.EffOpenRealInventory;
 import com.shanebeestudios.skbee.elements.other.effects.EffOpenSign;
@@ -86,6 +87,7 @@ import com.shanebeestudios.skbee.elements.other.expressions.ExprChatCompletions;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprChunkAt;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprChunkCoords;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprChunkInhabitedTime;
+import com.shanebeestudios.skbee.elements.other.expressions.ExprChunkKey;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprChunkLoadLevel;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprChunksWithinCuboid;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprClassInfoOf;
@@ -177,7 +179,9 @@ public class OtherElementRegistration {
         EffBlockRandomlyTick.register(registration);
         EffBlockstateUpdate.register(registration);
         EffBreakBlocksWithEffects.register(registration);
+        EffChunkLoad.register(registration);
         EffChunkRefresh.register(registration);
+        EffChunkUnload.register(registration);
         EffDispatchCommand.register(registration);
         EffEntityBlockStorage.register(registration);
         EffEntityDamageEntity.register(registration);
@@ -186,7 +190,6 @@ public class OtherElementRegistration {
         EffFreezePlayerConnection.register(registration);
         EffGiveOrDrop.register(registration);
         EffHurtAnimation.register(registration);
-        EffLoadChunk.register(registration);
         EffOpenContainerAnimation.register(registration);
         EffOpenRealInventory.register(registration);
         EffOpenSign.register(registration);
@@ -250,6 +253,7 @@ public class OtherElementRegistration {
         ExprChunkAt.register(registration);
         ExprChunkCoords.register(registration);
         ExprChunkInhabitedTime.register(registration);
+        ExprChunkKey.register(registration);
         ExprChunkLoadLevel.register(registration);
         ExprChunksWithinCuboid.register(registration);
         ExprClassInfoOf.register(registration);
